+++ src/deprecate.jl
@@ -1,2 +1,4 @@
-# example
-# @deprecate oldfun newfun
+@deprecate fadj(g) out_adjlist(g)
+@deprecate fadj(g, v) out_neighbors(g, v)
+@deprecate badj(g) in_adjlist(g)
+@deprecate badj(g, v) in_neighbors(g, v)