+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - Graph(dg) -> graph(dg)
 - remove fadj(g, v) in favor of out_neighbors(g, v)
 - more efficient equality test for DiGraph
+- remove fadj(g) in favor of out_adjlist